refactor(api): extract tree write operations to tree_writes.py
